third_party/blink/web_tests/platform/generic/accessibility/dl-role-expected.txt

This tests that AXRole for dl element is AXDescriptionList.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS dlElement.role is 'AXRole: AXDescriptionList'
PASS dtElement.role is 'AXRole: AXDescriptionListTerm'
PASS ddElement.role is 'AXRole: AXDescriptionListDetail'
PASS successfullyParsed is true

TEST COMPLETE

term
definition
